substantial refactoring for render/event handler methods and otherwise
