+++ src/classes/StringTable.ts
@@ -1,8 +1,10 @@
-import { StringState } from './StringState.js'
-import { StringStateCollection } from './StringStateCollection.js'
-import { StringManager } from './StringManager.js'
-import { Utilities } from '../static/Utilities.js'
-import { Note } from '../static/Note.js'
+import { StringState } from "./StringState.js"
+import { StringStateCollection } from "./StringStateCollection.js"
+import { StringSetEnum } from "../enums/StringSetEnum.js"
+import { TableRenders } from "../renders/TableRenders.js"
+import { TableHandlers } from "../events/TableEvents.js"
+import { Utilities } from "../static/Utilities.js"
+import { Note } from "../static/Note.js"
 
 export { StringTable }
 
@@ -12,14 +14,22 @@ export { StringTable }
 class StringTable {
     private _currentStrings: StringStateCollection
     private _stringCache: StringStateCollection
+    private _stringSetName: StringSetEnum | null = null
     private _canModifyGauge: boolean
     private _isCurrent: boolean
+
+    private _renders: TableRenders
+    private _handles: TableHandlers
 
     constructor(startingStrings: StringStateCollection) {
         this._currentStrings = startingStrings
         this._stringCache = new StringStateCollection()
         this._canModifyGauge = true
         this._isCurrent = false
+
+        this._renders = new TableRenders(this)
+        this._handles = new TableHandlers(this.renders)
+
     }
 
     public get currentStrings(): StringStateCollection {
@@ -38,6 +48,14 @@ class StringTable {
         this._stringCache = value
     }
 
+    public get stringSetName(): StringSetEnum | null {
+        return this._stringSetName
+    }
+
+    public set stringSetName(value: StringSetEnum | null) {
+        this._stringSetName = value
+    }
+
     public get canModifyGauge(): boolean {
         return this._canModifyGauge
     }
@@ -52,6 +70,22 @@ class StringTable {
 
     public set isCurrent(value: boolean) {
         this._isCurrent = value
+    }
+
+    public get renders(): TableRenders {
+        return this._renders
+    }
+
+    public set renders(value: TableRenders) {
+        this._renders = value
+    }
+
+    public get handles(): TableHandlers {
+        return this._handles
+    }
+
+    public set handles(value: TableHandlers) {
+        this._handles = value
     }
 
     /**
@@ -104,160 +138,6 @@ class StringTable {
     }
 
     /**
-     * Re-renders the guitar string table.
-     *
-     * @param {string} tableId The id for the string table.
-     */
-    public render(tableId: string) {
-        let strTable = document.createElement('table')
-        let tr = document.createElement('tr')
-
-        let tableHeaderCells = [
-            document.createElement('th'),
-            document.createElement('th'),
-            document.createElement('th'),
-            document.createElement('th'),
-            document.createElement('th'),
-            document.createElement('th'),
-        ]
-
-        tr.classList.add('row-top')
-        strTable.setAttribute('id', tableId)
-
-        tableHeaderCells[0].innerText = 'String'
-        tableHeaderCells[1].innerText = 'Note'
-        tableHeaderCells[2].innerText = 'Scale'
-        tableHeaderCells[3].innerText = 'Name'
-        tableHeaderCells[4].innerText = 'Gauge'
-        tableHeaderCells[5].innerText = 'Tension'
-
-        for (let headerCell of tableHeaderCells) {
-            tr.appendChild(headerCell)
-        }
-
-        strTable.appendChild(tr)
-
-        // Add string rows.
-        for (let i = 0; i < this.getNumStrings(); i++) {
-            let strRow = this.makeStringRow(i + 1, this.getString(i), this)
-
-            strTable.appendChild(strRow)
-        }
-
-        let tableElem = document.getElementById(tableId)!
-
-        tableElem.parentNode!.replaceChild(strTable, tableElem)
-    }
-
-    /**
-     * Handles conversions on change of scale length input.
-     * 
-     * @param scaleLengthBox 
-     * @param state 
-     */
-    public onChangeInputScaleLength(scaleLengthBox: any, state: StringState) {
-        scaleLengthBox.type = 'text'
-        scaleLengthBox.value = state.scaleLength.toString() + '"'
-
-        scaleLengthBox.onchange = (() => {
-            let inputScale = scaleLengthBox.value.trim()
-
-            // Temp: convert from mm to inches.
-            let convertFromMillimeters = inputScale.substring(inputScale.length - 2) === 'mm'
-
-            // Easter Egg: convert feet to inches if a single tick is input.
-            let convertFromFeet = inputScale.charAt(inputScale.length - 1) == "'"
-
-            // Trim out units from string.
-            if (inputScale.charAt(inputScale.length - 1) === '"' || convertFromFeet) {
-                inputScale = inputScale.substring(0, inputScale.length - 1)
-            }
-            else if (convertFromMillimeters) {
-                inputScale = inputScale.substring(0, inputScale.length - 2).trim()
-            }
-
-            // Attempt to convert to number.
-            inputScale = Number.parseFloat(inputScale)
-
-            /* 
-             * If it's a number, go ahead and set the scale and redraw the table.
-             * Otherwise, just return the original value.
-             */ 
-            if (typeof inputScale === 'number' && Number.isFinite(inputScale)) {
-                if (convertFromFeet) {
-                    inputScale *= 12
-                }
-                if (convertFromMillimeters) {
-                    // TODO: A smart rounding system? (i.e. only if the value is very close to a certain fraction of an inch (e.g., 1/4th, 1/8th))
-                    inputScale /= 25.4
-                }
-
-                state.scaleLength = inputScale
-                this.render('str-table')
-            }
-            else {
-                scaleLengthBox.value = state.scaleLength + '"'
-            }
-        }).bind(this)
-    }
-
-    /**
-     * Handles button click to pitch down.
-     * 
-     * @param buttonPitchDown 
-     * @param state 
-     */
-    public onClickButtonPitchDown(buttonPitchDown: HTMLInputElement, state: StringState) {
-        buttonPitchDown.onclick = (() => {
-            state.shiftPitch(-1)
-            this.render('str-table')
-        }).bind(this)
-    }
-
-    /**
-     * Handles button click to pitch up.
-     * 
-     * @param buttonPitchUp 
-     * @param state 
-     */
-    public onClickButtonPitchUp(buttonPitchUp: HTMLInputElement, state: StringState) {
-        buttonPitchUp.onclick = (() => {
-            state.shiftPitch(1)
-            this.render('str-table')
-        }).bind(this)
-    }
-
-    /**
-     * Handles button click to decrease gauge.
-     * 
-     * @param buttonGaugeDecrease
-     * @param state 
-     */
-    public onClickButtonDecreaseGauge(buttonGaugeDecrease: HTMLInputElement, state: StringState) {
-        buttonGaugeDecrease.onclick = (() => {
-            let currentSeries = StringManager.getInstance().getSeriesByBrandAndType(state.strInfo.brand, state.strInfo.type)
-
-            state.strInfo = currentSeries.getPreviousString(state.strInfo)
-            this.render('str-table')
-        }).bind(this)
-    }
-
-    /**
-     * Handles button click to increase gauge.
-     * 
-     * @param buttonGaugeIncrease 
-     * @param state 
-     */
-    public onClickButtonIncreaseGauge(buttonGaugeIncrease: HTMLInputElement, state: StringState) {
-        buttonGaugeIncrease.onclick = (() => {
-            let currentSeries = StringManager.getInstance().getSeriesByBrandAndType(state.strInfo.brand, state.strInfo.type)
-
-            state.strInfo = currentSeries.getNextString(state.strInfo)
-            this.render('str-table')
-        }).bind(this)
-    }
-
-    /**
      * Makes a row for a guitar string.
      *
      * @param {number} num The string number.
@@ -298,11 +178,11 @@ class StringTable {
         stringNum.appendChild(document.createTextNode(num.toString()))
 
         // Event handlers
-        this.onChangeInputScaleLength(scaleLengthBox, state)
-        this.onClickButtonPitchDown(buttonPitchDown, state)
-        this.onClickButtonPitchUp(buttonPitchUp, state)
-        this.onClickButtonDecreaseGauge(buttonDecreaseGauge, state)
-        this.onClickButtonIncreaseGauge(buttonIncreaseGauge, state)
+        this.handles.onChangeInputScaleLength(scaleLengthBox, state)
+        this.handles.onClickButtonPitchDown(buttonPitchDown, state)
+        this.handles.onClickButtonPitchUp(buttonPitchUp, state)
+        this.handles.onClickButtonDecreaseGauge(buttonDecreaseGauge, state)
+        this.handles.onClickButtonIncreaseGauge(buttonIncreaseGauge, state)
 
         // Adding each field to the row, as well as the `note-inner` element to the 'Note' field
         for (let field of fields) {
